Remove dynamic-loader scripts, simplify script loading
